pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zelscript"
version = "0.1.0"
description = "Small, single-purpose CLI helpers that plug into the wider Zel suite"
authors = [{name = "Zeldean", email = "deanvanzyl2004@gmail.com"}]
dependencies = ["click", "Pillow"]
requires-python = ">=3.8"

[project.scripts]
zelscript = "zelscript.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["zelscript*"]

[tool.setuptools.package-dir]
"" = "src"
